build: move structs
